requirements.txt
beautifulsoup4==4.13.3
discord.py==2.5.2
IMAPClient==3.0.1
python-dotenv==1.0.1
Requests==2.32.3
